Move collection of rule links to review-dependencies
