+++ lib/project-dependencies.js
@@ -1,5 +1,9 @@
 const ProjectJsonFiles = require('./project-json-files');
 
+module.exports = {
+  collect
+};
+
 async function collect(options, elmJson, reviewDirectDependencies, elmVersion) {
   const dependenciesEntries =
     elmJson.type === 'application'
@@ -68,63 +72,6 @@ I will try to review the project anyway, but you might get unexpected results…
   return projectDeps;
 }
 
-async function collectRuleLinks(reviewDirectDependencies, elmVersion) {
-  const elmJsonForReviewDependenciesPromises = Object.entries(
-    reviewDirectDependencies
-  )
-    .filter(([name]) => !name.startsWith('elm/'))
-    .map(([name, packageVersion]) =>
-      ProjectJsonFiles.getElmJsonFromElmHome(
-        elmVersion,
-        name,
-        packageVersion
-      ).catch(() => null)
-    );
-
-  const elmJsonForReviewDependencies = await Promise.all(
-    elmJsonForReviewDependenciesPromises
-  ).then((items) => items.filter(Boolean));
-
-  return computeLinksToRuleDocs(
-    reviewDirectDependencies,
-    elmJsonForReviewDependencies
-  );
-}
-
-function computeLinksToRuleDocs(
-  reviewDirectDependencies,
-  elmJsonForReviewDependencies
-) {
-  return Object.entries(reviewDirectDependencies).reduce(
-    (acc, [depName, packageVersion]) => {
-      const dep = elmJsonForReviewDependencies.find(
-        (elmJson) => elmJson.name === depName
-      );
-
-      if (!dep || !('jfmengels/elm-review' in dep.dependencies)) {
-        return acc;
-      }
-
-      const exposedModules = Array.isArray(dep['exposed-modules'])
-        ? dep['exposed-modules']
-        : Object.values(dep['exposed-modules']).reduce((acc, items) =>
-            acc.concat(items)
-          );
-
-      exposedModules.forEach((moduleName) => {
-        acc[moduleName] = linkToModule(depName, packageVersion, moduleName);
-      });
-      return acc;
-    },
-    {}
-  );
-}
-
-function linkToModule(dependencyName, packageVersion, moduleName) {
-  const urlModuleName = moduleName.split('.').join('-');
-  return `https://package.elm-lang.org/packages/${dependencyName}/${packageVersion}/${urlModuleName}`;
-}
-
 function defaultElmJson(name, packageVersion) {
   return {
     type: 'package',
@@ -141,8 +88,3 @@ function defaultElmJson(name, packageVersion) {
     'test-dependencies': {}
   };
 }
-
-module.exports = {
-  collect,
-  collectRuleLinks
-};
