+++ lib/runner.js
@@ -8,6 +8,7 @@ const AppWrapper = require('./app-wrapper');
 const ErrorMessage = require('./error-message');
 const {getProjectFiles} = require('./elm-files');
 const SuppressedErrors = require('./suppressed-errors');
+const ReviewDependencies = require('./review-dependencies');
 const ProjectDependencies = require('./project-dependencies');
 const {runReview, startReview, requestReview} = require('./run-review');
 
@@ -120,7 +121,7 @@ async function initializeApp(options, elmModulePath, reviewElmJson, appHash) {
 
   const suppressedErrorsP = SuppressedErrors.read(options);
 
-  const linksToRuleDocsP = ProjectDependencies.collectRuleLinks(
+  const linksToRuleDocsP = ReviewDependencies.collectRuleLinks(
     reviewElmJson.dependencies.direct,
     elmVersion
   );
